% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_measure.R
\name{balance_measure}
\alias{balance_measure}
\title{Balance measurement.}
\usage{
balance_measure(OS_ind, stage_list, data)
}
\arguments{
\item{OS_ind}{OS_ind is the survival indicator variable.}

\item{stage_list}{stage_list original of each scheme.}

\item{data}{Data set.}
}
\value{
Ranking of balance measurement and its standardized score.
}
\description{
\code{balance_measure} returns balance measurement for the grouping scheme.
}
\references{
Xu, W., et al. 'Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer.' Austin Biometrics and Biostatistics 2 (2015): 1014.
}
