% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_var_measure.R
\name{explain_var_measure}
\alias{explain_var_measure}
\title{Explained variation measurement.}
\usage{
explain_var_measure(main_list, stage_list, stage_list_2, covar_list, data)
}
\arguments{
\item{main_list}{main_list includes survival indicator variable, Duration time of survival variable and basic group variable.}

\item{stage_list}{stage_list original of each scheme.}

\item{stage_list_2}{stage_list_2 is numerical form of each scheme by using orther of stages information.}

\item{covar_list}{Covariate variables taking into consideration.}

\item{data}{Data set.}
}
\value{
Ranking of explained variation measurement and its standardized score.
}
\description{
\code{explain_var_measure} returns explained variation measurement for the grouping scheme.
}
\references{
Xu, W., et al. "Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer." Austin Biom Biostat 2 (2015): 1014.
}
