% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_cons_measure.R
\name{hz_cons_measure}
\alias{hz_cons_measure}
\title{Hazard consistency measurement.}
\usage{
hz_cons_measure(main_list, stage_list, covar_list, data)
}
\arguments{
\item{main_list}{main_list includes survival indicator variable, Duration time of survival variable and basic group variable.}

\item{stage_list}{stage_list is numerical form of each scheme by using other of stages information.}

\item{covar_list}{Covariate variables taking into consideration.}

\item{data}{Data set.}
}
\value{
Ranking of hazard consistency measurement and  standardized score.
}
\description{
\code{hz_cons_measure} returns Hazard consistency for the grouping scheme.
}
\references{
Xu, W., et al. 'Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer.' Austin Biometrics and Biostatistics 2 (2015): 1014.
}
