% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_dis_measure.R
\name{hz_dis_measure}
\alias{hz_dis_measure}
\title{Hazard discrimination measurement.}
\usage{
hz_dis_measure(main_list, stage_list, stage_list2, covar_list, data)
}
\arguments{
\item{main_list}{main_list includes survival indicator variable, Duration time of survival variable and basic group variable.}

\item{stage_list}{stage_list original of each scheme.}

\item{stage_list2}{stage_list2 is numerical form of each scheme by using other of stages information.}

\item{covar_list}{Covariate variables taking into consideration.}

\item{data}{Data set.}
}
\value{
Ranking of hazard discrimination measurement and its standardized score.
}
\description{
\code{hz_dis_measure} returns hazard discrimination for the grouping scheme.
}
\references{
Xu, W., et al. 'Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer.' Austin Biometrics and Biostatistics 2 (2015): 1014.
}
