% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_diff_measure.R
\name{lik_diff_measure}
\alias{lik_diff_measure}
\title{Likelihood difference measurement.}
\usage{
lik_diff_measure(main_list, stage_list, covar_list, data)
}
\arguments{
\item{main_list}{main_list includes survival indicator variable, Duration time of survival variable and basic group variable.}

\item{stage_list}{stage_list is numerical form of each scheme by using orther of stages information.}

\item{covar_list}{Covariate variables taking into consideration.}

\item{data}{Data set.}
}
\value{
Ranking of likelihood difference measurement and its standardized score.
}
\description{
\code{lik_diff_measure} returns likelihood difference for the grouping scheme.
}
\references{
Xu, W., et al. "Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer." Austin Biom Biostat 2 (2015): 1014.
}
