% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_rank.R
\name{overall_rank}
\alias{overall_rank}
\title{Overall Ranking.}
\usage{
overall_rank(HCM, HDM, LDM, EVM, BM, weight)
}
\arguments{
\item{HCM}{HCM is the hazard consistency measurement results.}

\item{HDM}{HDM is the hazard discrimination measurement results.}

\item{LDM}{LDM is the likelihood difference measurement results.}

\item{EVM}{EVM is the explained variance measurement results.}

\item{BM}{BM is the balance measurement results.}

\item{weight}{weight vector of five measurements.}
}
\value{
Overall score and overall ranking.
}
\description{
\code{overall_rank} returns overall ranking for the grouping scheme.
}
\references{
Xu, W., et al. 'Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer.' Austin Biometrics and Biostatistics 2 (2015): 1014.
}
