% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{Rdata}
\alias{Rdata}
\title{Simulation data for Genetic association models for X-chromosome SNPS}
\format{A data frame with 504 rows and 10 variables.}
\usage{
data(Rdata)
}
\description{
A simulated dataset containing 400 observations. The variables list as follows:
}
\details{
\itemize{
  \item Id  Identification number.
  \item Gender 1 for male, 0 for female. 
  \item Age Age variable.
  \item OS Survival indicator, 1 for death, 0 for censoring.
  \item survmonth Duration time of survival.
  \item Treatment Treament variable.
  \item Basic_group Basic group variable having the most number of stages.
  \item Scheme.1 Grouping scheme 1.
  \item Scheme.2 Grouping scheme 2.
  \item Scheme.3 Grouping scheme 3.
}
}
\keyword{datasets}
