% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remss.R
\docType{package}
\name{remss}
\alias{remss}
\alias{remss-package}
\title{remss: A package for refining evaluation methodology on stage system.}
\description{
TNM stage is important in treatment decision-making and outcome predicting. The existing Oropharyngeal Cancer                
(OPC) TNM stages have not made distinction of the two sub sites of HPV+ and HPV- diseases.We developed novel              
criteria to assess performance of the TNM stage grouping schemes based on parametric modeling adjusting on                    
important clinical factors. These criteria evaluate the TNM stage groupingscheme in five different measures:                  
hazard  consistency, hazard discrimination, explained variation, likelihood difference, and balance.
}
\section{xlink functions}{

rank, hz_cons_measure,hz_dis_measure,lik_diff_measure,explain_var_measure,balance_measure.
}

\references{
Xu, W., et al. "Refining evaluation methodology on TNM stage system: assessment on HPV-related oropharyngeal cancer." Austin Biom Biostat 2 (2015): 1014.
}
